src/requirements.txt
alembic
strawberry-graphql
SQLAlchemy
flask-jwt-extended
Flask
Flask-CORS
Flask-mail
Tornado
Werkzeug
psycopg2-binary
requests
strawberry-sqlalchemy-mapper